pyproject.toml
[project]
name = "paperbanana-v2"
version = "0.1.0"
description = "Academic figure generation with Agno agents"
requires-python = ">=3.12"
dependencies = [
    "agno",
    "chromadb",
    "dash",
    "google-genai",
    "pillow",
    "pydantic",
    "python-dotenv",
    "pyyaml",
]

[tool.setuptools.packages.find]
include = ["agents*", "retriever*"]
